refactor: register form
