+++ src/app/modules/auth/register.page/register.page.tsx
@@ -1,60 +1,10 @@
-import {
-  Button,
-  Checkbox,
-  InputField,
-  LinkButton,
-  SideImage,
-  Text,
-} from "@/app/atomic";
+import { LinkButton, SideImage, Text } from "@/app/atomic";
 import GuinaTeamImage from "../../../assets/svg/guina-team/Image.png";
 import { registerStrings } from "./register.strings";
 import { ArrowLeftIcon, Vector } from "@/app/assets/svg";
-import { FormProvider, useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { z } from "zod";
-
-const formSchema = z
-  .object({
-    name: z.string().nonempty({ message: "Insira seu nome completo" }),
-    email: z
-      .string()
-      .nonempty({ message: "Insira seu email" })
-      .toLowerCase()
-      .email({ message: "Email inválido" }),
-    password: z
-      .string()
-      .nonempty({ message: "Insira sua senha" })
-      .min(6, { message: "A senha deve conter no mínimo 6 dígitos" }),
-    confirmPassword: z
-      .string()
-      .nonempty({ message: "Confirme sua senha" })
-      .min(6, { message: "A senha deve conter no mínimo 6 dígitos" }),
-  })
-  .refine((data) => data.password === data.confirmPassword, {
-    message: "Senha incorreta",
-    path: ["confirmPassword"],
-  });
+import { RegisterForm } from "./components/register.form";
 
 export function RegisterPage() {
-  const methods = useForm({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      name: "",
-      email: "",
-      password: "",
-      confirmPassword: "",
-    },
-  });
-
-  const onSubmit = (formData: {
-    name: string;
-    email: string;
-    password: string;
-    confirmPassword: string;
-  }) => {
-    console.log("teste", formData);
-  };
-
   return (
     <div className="items-center justify-between h-screen flex flex-rol w-full">
       <div className="flex flex-1 flex-col h-full">
@@ -70,63 +20,7 @@ export function RegisterPage() {
               <Text variant="display">{registerStrings.registerTitle}</Text>
               <Text variant="body1">{registerStrings.registerSubtitle}</Text>
             </div>
-            <FormProvider {...methods}>
-              <form onSubmit={methods.handleSubmit(onSubmit)}>
-                <div className="flex flex-col items-start gap-md w-[400px]">
-                  <InputField
-                    name="name"
-                    label={registerStrings.nameInput.label}
-                    type="text"
-                    placeholder={registerStrings.nameInput.placeholder}
-                    className="flex w-full flex-col gap-sm"
-                  />
-                  <InputField
-                    name="email"
-                    label={registerStrings.emailInput.label}
-                    type="email"
-                    placeholder={registerStrings.emailInput.placeholder}
-                    className="flex w-full flex-col gap-sm"
-                  />
-                  <InputField
-                    name="password"
-                    label={registerStrings.passwordInput.label}
-                    type="password"
-                    placeholder={registerStrings.passwordInput.placeholder}
-                    className="flex w-full flex-col gap-sm"
-                  />
-                  <InputField
-                    name="confirmPassword"
-                    label={
-                      registerStrings.passwordInput.confirmationInput.label
-                    }
-                    type="password"
-                    placeholder={
-                      registerStrings.passwordInput.confirmationInput
-                        .placeholder
-                    }
-                    className="flex w-full flex-col gap-sm"
-                  />
-                  <div className="flex gap-x-sm">
-                    <Checkbox />
-                    <Text variant="body1">
-                      {registerStrings.readAndAgreed}
-                      <LinkButton variant="link" path="./">
-                        {" "}
-                        {registerStrings.termsOfUse}{" "}
-                      </LinkButton>
-                      {registerStrings.and}
-                      <LinkButton variant="link" path="./">
-                        {" "}
-                        {registerStrings.privacyPolicy}
-                      </LinkButton>
-                    </Text>
-                  </div>
-                  <Button type="submit" isLoading={false}>
-                    {registerStrings.ctaRegister}
-                  </Button>
-                </div>
-              </form>
-            </FormProvider>
+            <RegisterForm />
             <div className="flex flex-col gap-sm items-center">
               <div className="flex flex-rol items-center justify-center gap-xxs">
                 <Vector />
